見出し配下も辿れるように修正
